requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.0
    # via
    #   -r requirements.in
    #   django
click==8.0.1
    # via
    #   -r requirements.in
    #   pip-tools
compiler==0.2.0
    # via -r requirements.in
dj-database-url==0.5.0
    # via
    #   -r requirements.in
    #   django-heroku
django==3.2.4
    # via
    #   -r requirements.in
    #   django-heroku
    #   djangorestframework
django-heroku==0.3.1
    # via -r requirements.in
djangorestframework==3.11.1
    # via -r requirements.in
djangorestframework-jwt==1.11.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
django-apscheduler==0.6.2
    # via -r requirements.in
pep517==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
pip-tools==6.2.0
    # via -r requirements.in
psycopg2==2.8.5
    # via
    #   -r requirements.in
    #   django-heroku
psycopg2-binary==2.9.1
    # via -r requirements.in
pyjwt==1.7.1
    # via
    #   -r requirements.in
    #   djangorestframework-jwt
python-decouple==3.3
    # via -r requirements.in
pytz==2021.1
    # via
    #   -r requirements.in
    #   django
sqlparse==0.4.1
    # via
    #   -r requirements.in
    #   django
toml==0.10.2
    # via
    #   -r requirements.in
    #   pep517
wheel==0.36.2
    # via pip-tools
whitenoise==5.3.0
    # via
    #   -r requirements.in
    #   django-heroku

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools